feat(app): split cloud settings up in multiple pages
