+++ apps/app/src/app/types.ts
@@ -174,7 +174,10 @@ export type SettingsTab =
   | "ai"
   | "permissions"
   | "shell"
-  | "den"
+  | "cloud-account"
+  | "cloud-marketplaces"
+  | "cloud-workers"
+  | "cloud-providers"
   | "skills"
   | "extensions"
   | "environment"
