+++ apps/app/src/i18n/locales/ca.ts
@@ -984,6 +984,7 @@ export default {
   "settings.feedback_badge": "Llegim tots els missatges",
   "settings.feedback_desc": "Digueu-nos què us sembla bé i què us sembla dur. Els comentaris van directament a l'equip i ens ajuden a prioritzar el que s'enviarà a continuació.",
   "settings.feedback_title": "Ajuda a donar forma a OpenWork",
+  "settings.group_cloud": "Cloud",
   "settings.group_global": "Global",
   "settings.group_workspace": "Workspace",
   "settings.hide_titlebar": "Amaga la barra de títol",
@@ -1110,12 +1111,18 @@ export default {
   "settings.switch": "Canvia",
   "settings.tab_advanced": "Avançat",
   "settings.tab_appearance": "Aparença",
-  "settings.tab_cloud": "Cloud",
+  "settings.tab_cloud_account": "Account",
+  "settings.tab_cloud_marketplaces": "Marketplaces & Plugins",
+  "settings.tab_cloud_providers": "Cloud Providers",
+  "settings.tab_cloud_workers": "Cloud Workers",
   "settings.tab_debug": "Depuració",
   "settings.tab_description_advanced": "Inspecciona l'estat del runtime, l'estat de la connexió i els controls orientats a desenvolupador.",
   "settings.tab_description_appearance": "Ajusta l'aspecte d'OpenWork al desktop, al tema del sistema i al marc de l'app.",
   "settings.tab_description_debug": "Revisa els diagnòstics, els registres i les utilitats avançades de depuració del runtime.",
-  "settings.tab_description_den": "Gestiona la connexió a OpenWork Cloud, els workers allotjats i l'accés al workspace.",
+  "settings.tab_description_cloud_account": "Sign in, manage your organization, and configure your cloud connection.",
+  "settings.tab_description_cloud_marketplaces": "Browse and import plugins from your organization's marketplaces.",
+  "settings.tab_description_cloud_providers": "Import and manage LLM provider keys from your organization.",
+  "settings.tab_description_cloud_workers": "View and connect to cloud-hosted worker instances.",
   "settings.tab_description_environment": "Save API keys and tokens for local agents, skills, and MCP servers. Secrets stay on this device.",
   "settings.tab_description_extensions": "Gestiona apps MCP i Plugins d'OpenCode per a aquest workspace.",
   "settings.tab_description_general": "Connecta proveïdors, tria el model predeterminat, autoritza carpetes i controla el workspace OpenWork seleccionat juntament amb la seva connexió de runtime.",
