+++ apps/app/src/i18n/locales/en.ts
@@ -1023,6 +1023,7 @@ export default {
   "settings.feedback_badge": "We read every message",
   "settings.feedback_desc": "Tell us what feels great and what feels rough. Feedback goes straight to the team and helps us prioritize what ships next.",
   "settings.feedback_title": "Help shape OpenWork",
+  "settings.group_cloud": "Cloud",
   "settings.group_global": "Global",
   "settings.group_workspace": "Workspace",
   "settings.hide_titlebar": "Hide titlebar",
@@ -1180,12 +1181,18 @@ export default {
   "settings.switch": "Switch",
   "settings.tab_advanced": "Advanced",
   "settings.tab_appearance": "Appearance",
-  "settings.tab_cloud": "Cloud",
+  "settings.tab_cloud_account": "Account",
+  "settings.tab_cloud_marketplaces": "Marketplaces & Plugins",
+  "settings.tab_cloud_providers": "Cloud Providers",
+  "settings.tab_cloud_workers": "Cloud Workers",
   "settings.tab_debug": "Debug",
   "settings.tab_description_advanced": "Inspect runtime health, connection state, and developer-facing controls.",
   "settings.tab_description_appearance": "Adjust how OpenWork looks across desktop, system theme, and app chrome.",
+  "settings.tab_description_cloud_account": "Sign in, manage your organization, and configure your cloud connection.",
+  "settings.tab_description_cloud_marketplaces": "Browse and import plugins from your organization's marketplaces.",
+  "settings.tab_description_cloud_providers": "Import and manage LLM provider keys from your organization.",
+  "settings.tab_description_cloud_workers": "View and connect to cloud-hosted worker instances.",
   "settings.tab_description_debug": "Review runtime diagnostics, logs, and low-level debugging utilities.",
-  "settings.tab_description_den": "Manage your OpenWork Cloud connection, hosted workers, and workspace access.",
   "settings.tab_description_extensions": "Manage MCP apps and OpenCode plugins for this workspace.",
   "settings.tab_description_general": "Connect providers, choose the default model, authorize folders, and control the selected OpenWork workspace plus its runtime connection.",
   "settings.environment.add_button": "Add variable",
