+++ apps/app/src/i18n/locales/es.ts
@@ -984,6 +984,7 @@ export default {
   "settings.feedback_badge": "Leemos cada mensaje",
   "settings.feedback_desc": "Cuéntanos qué se siente genial y qué se siente difícil. Tus comentarios van directamente al equipo y nos ayudan a priorizar lo próximo que lancemos.",
   "settings.feedback_title": "Ayuda a dar forma a OpenWork",
+  "settings.group_cloud": "Cloud",
   "settings.group_global": "Global",
   "settings.group_workspace": "Espacio de trabajo",
   "settings.hide_titlebar": "Ocultar barra de título",
@@ -1110,12 +1111,18 @@ export default {
   "settings.switch": "Cambiar",
   "settings.tab_advanced": "Avanzado",
   "settings.tab_appearance": "Apariencia",
-  "settings.tab_cloud": "Cloud",
+  "settings.tab_cloud_account": "Account",
+  "settings.tab_cloud_marketplaces": "Marketplaces & Plugins",
+  "settings.tab_cloud_providers": "Cloud Providers",
+  "settings.tab_cloud_workers": "Cloud Workers",
   "settings.tab_debug": "Depurar",
   "settings.tab_description_advanced": "Inspecciona el estado del runtime, el estado de la conexión y los controles orientados a desarrolladores.",
   "settings.tab_description_appearance": "Ajusta la apariencia de OpenWork en el escritorio, el tema del sistema y la app de Chrome.",
   "settings.tab_description_debug": "Revisa diagnósticos, registros y utilidades avanzadas de depuración del runtime.",
-  "settings.tab_description_den": "Gestiona tu conexión con OpenWork Cloud y tu acceso alojado a workers y espacios de trabajo.",
+  "settings.tab_description_cloud_account": "Sign in, manage your organization, and configure your cloud connection.",
+  "settings.tab_description_cloud_marketplaces": "Browse and import plugins from your organization's marketplaces.",
+  "settings.tab_description_cloud_providers": "Import and manage LLM provider keys from your organization.",
+  "settings.tab_description_cloud_workers": "View and connect to cloud-hosted worker instances.",
   "settings.tab_description_environment": "Save API keys and tokens for local agents, skills, and MCP servers. Secrets stay on this device.",
   "settings.tab_description_extensions": "Gestiona aplicaciones MCP y Plugins de OpenCode para este espacio de trabajo.",
   "settings.tab_description_general": "Conecta proveedores, elige el modelo predeterminado, autoriza carpetas y controla el espacio de trabajo seleccionado junto con su conexión de runtime.",
