+++ apps/app/src/i18n/locales/fr.ts
@@ -984,6 +984,7 @@ export default {
   "settings.feedback_badge": "Nous lisons chaque message",
   "settings.feedback_desc": "Dites-nous ce qui vous plaît et ce qui vous semble difficile. Les retours vont directement à l'équipe et nous aident à prioriser ce qui sera livré ensuite.",
   "settings.feedback_title": "Aidez à façonner OpenWork",
+  "settings.group_cloud": "Cloud",
   "settings.group_global": "Global",
   "settings.group_workspace": "Espace de travail",
   "settings.hide_titlebar": "Masquer la barre de titre",
@@ -1110,12 +1111,18 @@ export default {
   "settings.switch": "Basculer",
   "settings.tab_advanced": "Avancé",
   "settings.tab_appearance": "Apparence",
-  "settings.tab_cloud": "Cloud",
+  "settings.tab_cloud_account": "Account",
+  "settings.tab_cloud_marketplaces": "Marketplaces & Plugins",
+  "settings.tab_cloud_providers": "Cloud Providers",
+  "settings.tab_cloud_workers": "Cloud Workers",
   "settings.tab_debug": "Débogage",
   "settings.tab_description_advanced": "Inspectez la santé du runtime, l'état de connexion et les contrôles destinés aux développeurs.",
   "settings.tab_description_appearance": "Ajustez l'apparence d'OpenWork sur desktop, thème système et chrome de l'application.",
   "settings.tab_description_debug": "Consultez les diagnostics du runtime, les journaux et les utilitaires de débogage bas niveau.",
-  "settings.tab_description_den": "Gérez votre connexion OpenWork Cloud, les workers hébergés et l'accès à l'espace de travail.",
+  "settings.tab_description_cloud_account": "Sign in, manage your organization, and configure your cloud connection.",
+  "settings.tab_description_cloud_marketplaces": "Browse and import plugins from your organization's marketplaces.",
+  "settings.tab_description_cloud_providers": "Import and manage LLM provider keys from your organization.",
+  "settings.tab_description_cloud_workers": "View and connect to cloud-hosted worker instances.",
   "settings.tab_description_environment": "Save API keys and tokens for local agents, skills, and MCP servers. Secrets stay on this device.",
   "settings.tab_description_extensions": "Gérez les applications MCP et les Plugins OpenCode pour cet espace de travail.",
   "settings.tab_description_general": "Connectez des fournisseurs, choisissez le modèle par défaut, autorisez des dossiers et contrôlez l'espace de travail OpenWork sélectionné ainsi que sa connexion runtime.",
