+++ apps/app/src/i18n/locales/ja.ts
@@ -960,6 +960,7 @@ export default {
   "settings.feedback_badge": "すべてのメッセージを読んでいます",
   "settings.feedback_desc": "良い点や改善点を教えてください。フィードバックはチームに直接届き、次に何をリリースするかの優先順位付けに役立ちます。",
   "settings.feedback_title": "OpenWorkを改善する",
+  "settings.group_cloud": "Cloud",
   "settings.group_global": "グローバル",
   "settings.group_workspace": "ワークスペース",
   "settings.hide_titlebar": "タイトルバーを非表示",
@@ -1086,12 +1087,18 @@ export default {
   "settings.switch": "切り替え",
   "settings.tab_advanced": "詳細",
   "settings.tab_appearance": "外観",
-  "settings.tab_cloud": "クラウド",
+  "settings.tab_cloud_account": "Account",
+  "settings.tab_cloud_marketplaces": "Marketplaces & Plugins",
+  "settings.tab_cloud_providers": "Cloud Providers",
+  "settings.tab_cloud_workers": "Cloud Workers",
   "settings.tab_debug": "デバッグ",
   "settings.tab_description_advanced": "初期設定をやり直すためにOpenWorkのローカル状態をリセットします。",
   "settings.tab_description_appearance": "テーマとユーザーインターフェースの外観をカスタマイズします。",
   "settings.tab_description_debug": "ランタイムダイアログ、ログ、低レベルデバッグユーティリティを確認します。",
-  "settings.tab_description_den": "OpenWork Cloud接続、ホストワーカー、ワークスペースアクセスを管理します。",
+  "settings.tab_description_cloud_account": "Sign in, manage your organization, and configure your cloud connection.",
+  "settings.tab_description_cloud_marketplaces": "Browse and import plugins from your organization's marketplaces.",
+  "settings.tab_description_cloud_providers": "Import and manage LLM provider keys from your organization.",
+  "settings.tab_description_cloud_workers": "View and connect to cloud-hosted worker instances.",
   "settings.tab_description_extensions": "このワークスペースのMCPアプリとOpenCodeプラグインを管理します。",
   "settings.tab_description_general": "プロバイダーを接続し、デフォルトモデルを選択し、フォルダへのアクセスを許可し、OpenWorkワークスペースとそのランタイム接続を管理します。",
   "settings.tab_description_environment": "ローカルのエージェント、skills、MCP servers が使う API キーやトークンを保存します。シークレットはこのデバイスに保持されます。",
