+++ apps/app/src/i18n/locales/pt-BR.ts
@@ -966,6 +966,7 @@ export default {
   "settings.feedback_badge": "Lemos cada mensagem",
   "settings.feedback_desc": "Conte-nos o que está ótimo e o que está difícil. O feedback vai direto para a equipe e nos ajuda a priorizar o que lançamos a seguir.",
   "settings.feedback_title": "Ajude a moldar o OpenWork",
+  "settings.group_cloud": "Cloud",
   "settings.group_global": "Global",
   "settings.group_workspace": "Workspace",
   "settings.hide_titlebar": "Ocultar barra de título",
@@ -1092,12 +1093,18 @@ export default {
   "settings.switch": "Alternar",
   "settings.tab_advanced": "Avançado",
   "settings.tab_appearance": "Aparência",
-  "settings.tab_cloud": "Cloud",
+  "settings.tab_cloud_account": "Account",
+  "settings.tab_cloud_marketplaces": "Marketplaces & Plugins",
+  "settings.tab_cloud_providers": "Cloud Providers",
+  "settings.tab_cloud_workers": "Cloud Workers",
   "settings.tab_debug": "Depuração",
   "settings.tab_description_advanced": "Inspecione a saúde do runtime, estado da conexão e controles para desenvolvedores.",
   "settings.tab_description_appearance": "Ajuste a aparência do OpenWork no desktop, tema do sistema e decoração do app.",
   "settings.tab_description_debug": "Revise diagnósticos do runtime, logs e utilitários de depuração de baixo nível.",
-  "settings.tab_description_den": "Gerencie sua conexão OpenWork Cloud, workers hospedados e acesso a workspaces.",
+  "settings.tab_description_cloud_account": "Sign in, manage your organization, and configure your cloud connection.",
+  "settings.tab_description_cloud_marketplaces": "Browse and import plugins from your organization's marketplaces.",
+  "settings.tab_description_cloud_providers": "Import and manage LLM provider keys from your organization.",
+  "settings.tab_description_cloud_workers": "View and connect to cloud-hosted worker instances.",
   "settings.tab_description_environment": "Save API keys and tokens for local agents, skills, and MCP servers. Secrets stay on this device.",
   "settings.tab_description_extensions": "Gerencie apps MCP e plugins OpenCode para este workspace.",
   "settings.tab_description_general": "Conecte provedores, escolha o modelo padrão, autorize pastas e controle o workspace OpenWork selecionado e sua conexão de runtime.",
