+++ apps/app/src/i18n/locales/ru.ts
@@ -1105,6 +1105,7 @@ export default {
   "settings.feedback_badge": "Мы читаем каждое сообщение",
   "settings.feedback_desc": "Расскажите, что вам нравится, а что вызывает трудности. Отзыв сразу попадает команде и помогает понять, что выпускать дальше.",
   "settings.feedback_title": "Помогите сформировать OpenWork",
+  "settings.group_cloud": "Cloud",
   "settings.group_global": "Глобально",
   "settings.group_workspace": "Рабочее пространство",
   "settings.hide_titlebar": "Скрыть title bar",
@@ -1257,12 +1258,18 @@ export default {
   "settings.switch": "Переключить",
   "settings.tab_advanced": "Дополнительно",
   "settings.tab_appearance": "Внешний вид",
-  "settings.tab_cloud": "Cloud",
+  "settings.tab_cloud_account": "Account",
+  "settings.tab_cloud_marketplaces": "Marketplaces & Plugins",
+  "settings.tab_cloud_providers": "Cloud Providers",
+  "settings.tab_cloud_workers": "Cloud Workers",
   "settings.tab_debug": "Отладка",
   "settings.tab_description_advanced": "Проверяйте состояние среды выполнения, состояние подключения и инструменты для разработчика.",
   "settings.tab_description_appearance": "Настройте внешний вид OpenWork на уровне desktop, system theme и chrome приложения.",
   "settings.tab_description_debug": "Просматривайте диагностику среды выполнения, журналы и низкоуровневые инструменты отладки.",
-  "settings.tab_description_den": "Управляйте подключением OpenWork Cloud, hosted workers и доступом к workspace.",
+  "settings.tab_description_cloud_account": "Sign in, manage your organization, and configure your cloud connection.",
+  "settings.tab_description_cloud_marketplaces": "Browse and import plugins from your organization's marketplaces.",
+  "settings.tab_description_cloud_providers": "Import and manage LLM provider keys from your organization.",
+  "settings.tab_description_cloud_workers": "View and connect to cloud-hosted worker instances.",
   "settings.tab_description_extensions": "Управляйте MCP apps и OpenCode plugins для этого workspace.",
   "settings.tab_description_general": "Подключайте провайдеров, выбирайте модель по умолчанию, авторизуйте папки и управляйте выбранным workspace OpenWork вместе с его подключением к среде выполнения.",
   "settings.environment.add_button": "Добавить переменную",
