+++ apps/app/src/i18n/locales/th.ts
@@ -961,6 +961,7 @@ export default {
   "settings.feedback_badge": "เราอ่านทุกข้อความ",
   "settings.feedback_desc": "บอกสิ่งที่ชอบและสิ่งที่ยังไม่ดีพอ ความคิดเห็นส่งตรงถึงทีมและช่วยให้เรารู้ว่าจะพัฒนาอะไรต่อไป",
   "settings.feedback_title": "ช่วยพัฒนา OpenWork",
+  "settings.group_cloud": "Cloud",
   "settings.group_global": "ทั่วไป",
   "settings.group_workspace": "พื้นที่ทำงาน",
   "settings.hide_titlebar": "ซ่อนแถบชื่อหน้าต่าง",
@@ -1087,12 +1088,18 @@ export default {
   "settings.switch": "สลับ",
   "settings.tab_advanced": "ขั้นสูง",
   "settings.tab_appearance": "ธีม",
-  "settings.tab_cloud": "Cloud",
+  "settings.tab_cloud_account": "Account",
+  "settings.tab_cloud_marketplaces": "Marketplaces & Plugins",
+  "settings.tab_cloud_providers": "Cloud Providers",
+  "settings.tab_cloud_workers": "Cloud Workers",
   "settings.tab_debug": "ดีบัก",
   "settings.tab_description_advanced": "ตรวจสอบสถานะรันไทม์ การเชื่อมต่อ และตัวควบคุมสำหรับนักพัฒนา",
   "settings.tab_description_appearance": "ปรับรูปลักษณ์ OpenWork ทั้งเดสก์ท็อป ธีมระบบ และ app chrome",
   "settings.tab_description_debug": "ตรวจสอบข้อมูลวินิจฉัยรันไทม์ logs และเครื่องมือดีบักระดับต่ำ",
-  "settings.tab_description_den": "จัดการการเชื่อมต่อ OpenWork Cloud, workers ที่โฮสต์ และการเข้าถึงพื้นที่ทำงาน",
+  "settings.tab_description_cloud_account": "Sign in, manage your organization, and configure your cloud connection.",
+  "settings.tab_description_cloud_marketplaces": "Browse and import plugins from your organization's marketplaces.",
+  "settings.tab_description_cloud_providers": "Import and manage LLM provider keys from your organization.",
+  "settings.tab_description_cloud_workers": "View and connect to cloud-hosted worker instances.",
   "settings.tab_description_environment": "Save API keys and tokens for local agents, skills, and MCP servers. Secrets stay on this device.",
   "settings.tab_description_extensions": "จัดการแอป MCP และ OpenCode plugins สำหรับพื้นที่ทำงานนี้",
   "settings.tab_description_general": "เชื่อมต่อผู้ให้บริการ เลือกโมเดลเริ่มต้น อนุญาตโฟลเดอร์ และควบคุมพื้นที่ทำงาน OpenWork ที่เลือกพร้อมการเชื่อมต่อรันไทม์",
