+++ apps/app/src/i18n/locales/vi.ts
@@ -961,6 +961,7 @@ export default {
   "settings.feedback_badge": "Chúng tôi đọc từng tin nhắn",
   "settings.feedback_desc": "Cho chúng tôi biết điều gì tuyệt vời và điều gì chưa ổn. Phản hồi sẽ được gửi thẳng đến nhóm và giúp chúng tôi ưu tiên những gì phát triển tiếp theo.",
   "settings.feedback_title": "Góp phần định hình OpenWork",
+  "settings.group_cloud": "Cloud",
   "settings.group_global": "Toàn cục",
   "settings.group_workspace": "Workspace",
   "settings.hide_titlebar": "Ẩn thanh tiêu đề",
@@ -1087,12 +1088,18 @@ export default {
   "settings.switch": "Chuyển",
   "settings.tab_advanced": "Nâng cao",
   "settings.tab_appearance": "Giao diện",
-  "settings.tab_cloud": "Cloud",
+  "settings.tab_cloud_account": "Account",
+  "settings.tab_cloud_marketplaces": "Marketplaces & Plugins",
+  "settings.tab_cloud_providers": "Cloud Providers",
+  "settings.tab_cloud_workers": "Cloud Workers",
   "settings.tab_debug": "Gỡ lỗi",
   "settings.tab_description_advanced": "Kiểm tra trạng thái runtime, kết nối và điều khiển dành cho nhà phát triển.",
   "settings.tab_description_appearance": "Điều chỉnh giao diện OpenWork trên desktop, giao diện hệ thống và chrome ứng dụng.",
   "settings.tab_description_debug": "Xem chẩn đoán runtime, nhật ký và tiện ích gỡ lỗi cấp thấp.",
-  "settings.tab_description_den": "Quản lý kết nối OpenWork Cloud, worker được lưu trữ và quyền truy cập workspace.",
+  "settings.tab_description_cloud_account": "Sign in, manage your organization, and configure your cloud connection.",
+  "settings.tab_description_cloud_marketplaces": "Browse and import plugins from your organization's marketplaces.",
+  "settings.tab_description_cloud_providers": "Import and manage LLM provider keys from your organization.",
+  "settings.tab_description_cloud_workers": "View and connect to cloud-hosted worker instances.",
   "settings.tab_description_environment": "Save API keys and tokens for local agents, skills, and MCP servers. Secrets stay on this device.",
   "settings.tab_description_extensions": "Quản lý ứng dụng MCP và plugins OpenCode cho workspace này.",
   "settings.tab_description_general": "Kết nối provider, chọn model mặc định, cấp quyền thư mục và kiểm soát workspace OpenWork cùng kết nối runtime.",
