+++ apps/app/src/i18n/locales/zh.ts
@@ -964,6 +964,7 @@ export default {
   "settings.feedback_badge": "我们阅读每条反馈",
   "settings.feedback_desc": "告诉我们哪里体验流畅、哪里需要改进。反馈直达团队，帮助我们确定下一步优先级。",
   "settings.feedback_title": "帮助塑造OpenWork",
+  "settings.group_cloud": "Cloud",
   "settings.group_global": "全局",
   "settings.group_workspace": "工作区",
   "settings.hide_titlebar": "隐藏标题栏",
@@ -1090,12 +1091,18 @@ export default {
   "settings.switch": "切换",
   "settings.tab_advanced": "高级",
   "settings.tab_appearance": "外观",
-  "settings.tab_cloud": "Cloud",
+  "settings.tab_cloud_account": "Account",
+  "settings.tab_cloud_marketplaces": "Marketplaces & Plugins",
+  "settings.tab_cloud_providers": "Cloud Providers",
+  "settings.tab_cloud_workers": "Cloud Workers",
   "settings.tab_debug": "调试",
   "settings.tab_description_advanced": "检查运行时健康状态、连接状态和面向开发者的控制项。",
   "settings.tab_description_appearance": "调整OpenWork在桌面、系统主题和应用外观方面的显示效果。",
   "settings.tab_description_debug": "查看运行时诊断信息、日志和底层调试工具。",
-  "settings.tab_description_den": "管理OpenWork Cloud连接、托管工作区和工作区访问。",
+  "settings.tab_description_cloud_account": "Sign in, manage your organization, and configure your cloud connection.",
+  "settings.tab_description_cloud_marketplaces": "Browse and import plugins from your organization's marketplaces.",
+  "settings.tab_description_cloud_providers": "Import and manage LLM provider keys from your organization.",
+  "settings.tab_description_cloud_workers": "View and connect to cloud-hosted worker instances.",
   "settings.tab_description_extensions": "管理此工作区的MCP应用和OpenCode插件。",
   "settings.tab_description_general": "连接提供商、选择默认模型、授权文件夹，以及控制所选OpenWork工作区和运行时连接。",
   "settings.tab_description_environment": "保存本机 agents、skills 和 MCP servers 使用的 API keys 与 tokens。Secret 只保留在这台设备上。",
