+++ apps/app/src/react-app/domains/settings/shell/settings-page.tsx
@@ -4,15 +4,18 @@ import {
   ArrowLeft,
   Bug,
   ChevronDown,
-  Cloud,
+  CloudCog,
   Cog,
+  Container,
   FolderLock,
   Paintbrush,
   Puzzle,
   RefreshCcw,
   ShieldCheck,
   Sparkles,
+  Store,
   Terminal,
+  UserCircle,
   Wrench,
 } from "lucide-react";
 
@@ -57,8 +60,14 @@ export function getSettingsTabIcon(tab: SettingsTab) {
       return Paintbrush;
     case "permissions":
       return FolderLock;
-    case "den":
-      return Cloud;
+    case "cloud-account":
+      return UserCircle;
+    case "cloud-marketplaces":
+      return Store;
+    case "cloud-workers":
+      return Container;
+    case "cloud-providers":
+      return CloudCog;
     case "skills":
       return Sparkles;
     case "extensions":
@@ -88,8 +97,14 @@ export function getSettingsTabLabel(tab: SettingsTab) {
       return "Customization";
     case "permissions":
       return "Permissions";
-    case "den":
-      return t("settings.tab_cloud");
+    case "cloud-account":
+      return t("settings.tab_cloud_account");
+    case "cloud-marketplaces":
+      return t("settings.tab_cloud_marketplaces");
+    case "cloud-workers":
+      return t("settings.tab_cloud_workers");
+    case "cloud-providers":
+      return t("settings.tab_cloud_providers");
     case "skills":
       return t("settings.tab_skills");
     case "extensions":
@@ -121,8 +136,14 @@ export function getSettingsTabDescription(tab: SettingsTab) {
       return "Branding, visibility, and shell controls";
     case "permissions":
       return "Authorized folders and file access";
-    case "den":
-      return t("settings.tab_description_den");
+    case "cloud-account":
+      return t("settings.tab_description_cloud_account");
+    case "cloud-marketplaces":
+      return t("settings.tab_description_cloud_marketplaces");
+    case "cloud-workers":
+      return t("settings.tab_description_cloud_workers");
+    case "cloud-providers":
+      return t("settings.tab_description_cloud_providers");
     case "skills":
       return t("settings.tab_description_skills");
     case "extensions":
@@ -151,10 +172,17 @@ export function getWorkspaceSettingsTabs(): SettingsTab[] {
 }
 
 export function getGlobalSettingsTabs(developerMode: boolean): SettingsTab[] {
-  const tabs: SettingsTab[] = ["ai", "shell", "den", "appearance", "environment", "updates", "recovery"];
+  const tabs: SettingsTab[] = ["ai", "shell", "appearance", "environment", "updates", "recovery"];
   if (developerMode) tabs.push("debug");
   return tabs;
 }
+
+export const CLOUD_SETTINGS_TABS: SettingsTab[] = [
+  "cloud-account",
+  "cloud-marketplaces",
+  "cloud-workers",
+  "cloud-providers",
+];
 
 type SettingsPageProps = {
   activeTab: SettingsTab;
@@ -184,6 +212,7 @@ type SettingsSidebarProps = Pick<SettingsPageProps, "activeTab" | "onSelectTab" 
 export function SettingsSidebar(props: SettingsSidebarProps) {
   const workspaceTabs = getWorkspaceSettingsTabs();
   const globalTabs = getGlobalSettingsTabs(props.developerMode);
+  const cloudTabs = CLOUD_SETTINGS_TABS;
 
   return (
     <Sidebar className="mac:**:data-[sidebar=sidebar]:bg-transparent">
@@ -270,6 +299,29 @@ export function SettingsSidebar(props: SettingsSidebarProps) {
           <SidebarGroupContent>
             <SidebarMenu>
               {globalTabs.map((tab) => {
+                const Icon = getSettingsTabIcon(tab);
+                return (
+                  <SidebarMenuItem key={tab}>
+                    <SidebarMenuButton
+                      type="button"
+                      isActive={props.activeTab === tab}
+                      onClick={() => props.onSelectTab(tab)}
+                    >
+                      <Icon />
+                      <span>{getSettingsTabLabel(tab)}</span>
+                    </SidebarMenuButton>
+                  </SidebarMenuItem>
+                );
+              })}
+            </SidebarMenu>
+          </SidebarGroupContent>
+        </SidebarGroup>
+
+        <SidebarGroup>
+          <SidebarGroupLabel>{t("settings.group_cloud")}</SidebarGroupLabel>
+          <SidebarGroupContent>
+            <SidebarMenu>
+              {cloudTabs.map((tab) => {
                 const Icon = getSettingsTabIcon(tab);
                 return (
                   <SidebarMenuItem key={tab}>
