+++ apps/app/src/react-app/shell/settings-route.tsx
@@ -37,8 +37,11 @@ import { AuthorizedFoldersPanel } from "../domains/settings/panels/authorized-fo
 import { SettingsStack } from "../domains/settings/settings-section";
 import { AdvancedView } from "../domains/settings/pages/advanced-view";
 import { AppearanceView } from "../domains/settings/pages/appearance-view";
+import { CloudAccountView } from "../domains/settings/pages/cloud-account-view";
+import { CloudMarketplacesView } from "../domains/settings/pages/cloud-marketplaces-view";
+import { CloudProvidersView } from "../domains/settings/pages/cloud-providers-view";
+import { CloudWorkersView } from "../domains/settings/pages/cloud-workers-view";
 import { DebugView } from "../domains/settings/pages/debug-view";
-import { DenView } from "../domains/settings/pages/den-view";
 import { EnvironmentView } from "../domains/settings/pages/environment-view";
 import { ExtensionsView } from "../domains/settings/pages/extensions-view";
 import { McpView } from "../domains/settings/pages/mcp-view";
@@ -49,6 +52,8 @@ import { UpdatesView } from "../domains/settings/pages/updates-view";
 import { useDebugViewModel } from "../domains/settings/state/debug-view-model";
 import { useMessagingViewProps } from "../domains/settings/state/messaging-view-state";
 import { useElectronUpdaterState } from "../domains/settings/state/electron-updater-state";
+import { CloudSessionProvider } from "../domains/settings/cloud/cloud-session-provider";
+import { useDenSession } from "../domains/settings/cloud/use-den-session";
 import { useBootState } from "./boot-state";
 import { SettingsShell } from "../domains/settings/shell/settings-shell";
 import { createExtensionsStore, useExtensionsStoreSnapshot } from "../domains/settings/state/extensions-store";
@@ -283,7 +288,6 @@ function parseSettingsPath(pathname: string): {
     case "ai":
     case "permissions":
     case "shell":
-    case "den":
     case "skills":
     case "advanced":
     case "appearance":
@@ -292,6 +296,13 @@ function parseSettingsPath(pathname: string): {
     case "recovery":
     case "debug":
       return { tab: head, redirectPath: null };
+    case "cloud-account":
+    case "cloud-marketplaces":
+    case "cloud-workers":
+    case "cloud-providers":
+      return { tab: head, redirectPath: null };
+    case "den":
+      return { tab: "cloud-account", redirectPath: "cloud-account" };
     case "extensions":
       if (tail === "mcp") return { tab: "extensions", redirectPath: null, extensionsSection: "mcp" };
       if (tail === "skills") return { tab: "extensions", redirectPath: null, extensionsSection: "skills" };
@@ -369,7 +380,7 @@ function applyThemeMode(mode: PersistedThemeMode) {
   document.documentElement.dataset.theme = resolved;
 }
 
-export function SettingsRoute() {
+function SettingsRouteContent() {
   const navigate = useNavigate();
   const location = useLocation();
   const params = useParams<{ workspaceId?: string }>();
@@ -679,6 +690,11 @@ export function SettingsRoute() {
   const connectionsSnapshot = useConnectionsStoreSnapshot(connectionsStore);
   const providerAuthSnapshot = useProviderAuthStoreSnapshot(providerAuthStore);
   useExtensionsStoreSnapshot(extensionsStore);
+
+  const denSession = useDenSession({
+    developerMode,
+    openLink: (url) => platform.openLink(url),
+  });
 
   const shareWorkspaceState = useShareWorkspaceState({
     workspaces,
@@ -1100,7 +1116,7 @@ export function SettingsRoute() {
   useCloudProviderAutoSync(providerAuthStore.runCloudProviderSync);
 
   useEffect(() => {
-    if (route.tab !== "den") return;
+    if (route.tab !== "cloud-providers") return;
     void providerAuthStore.runCloudProviderSync("settings_cloud_opened");
   }, [providerAuthStore, route.tab]);
 
@@ -1437,6 +1453,10 @@ export function SettingsRoute() {
     return <Navigate to={workspaceSettingsRoute(selectedWorkspaceId, settingsPathForRoute(route))} replace state={location.state} />;
   }
 
+  const openCloudAccountSettings = () => {
+    navigate(selectedWorkspaceId ? workspaceSettingsRoute(selectedWorkspaceId, "cloud-account") : "/settings/cloud-account");
+  };
+
   const settingsView = (() => {
     switch (route.tab) {
       case "general":
@@ -1602,18 +1622,38 @@ export function SettingsRoute() {
 
           />
         );
-      case "den":
+      case "cloud-account":
         return (
-          <DenView
+          <CloudAccountView
             developerMode={developerMode}
+            session={denSession}
+          />
+        );
+      case "cloud-marketplaces":
+        return (
+          <CloudMarketplacesView
             extensions={extensionsStore}
-            openLink={(url) => platform.openLink(url)}
+            session={denSession}
+            onOpenAccount={openCloudAccountSettings}
+          />
+        );
+      case "cloud-workers":
+        return (
+          <CloudWorkersView
             connectRemoteWorkspace={async () => false}
+            onOpenAccount={openCloudAccountSettings}
+          />
+        );
+      case "cloud-providers":
+        return (
+          <CloudProvidersView
             cloudOrgProviders={providerAuthSnapshot.cloudOrgProviders}
-            importedCloudProviders={providerAuthSnapshot.importedCloudProviders}
-            refreshCloudOrgProviders={providerAuthStore.refreshCloudOrgProviders}
             connectCloudProvider={providerAuthStore.connectCloudProvider}
+            importedCloudProviders={providerAuthSnapshot.importedCloudProviders}
+            onOpenAccount={openCloudAccountSettings}
+            refreshCloudOrgProviders={providerAuthStore.refreshCloudOrgProviders}
             removeCloudProvider={providerAuthStore.removeCloudProvider}
+            session={denSession}
           />
         );
       case "advanced":
@@ -1898,5 +1938,13 @@ export function SettingsRoute() {
         onClose={() => setModelPickerOpen(false)}
       />
     </>
+  );
+}
+
+export function SettingsRoute() {
+  return (
+    <CloudSessionProvider>
+      <SettingsRouteContent />
+    </CloudSessionProvider>
   );
 }
